pyproject.toml
[project]
name = "afm-net-short-scraper"
version = "0.1.0"
description = "Scraper for AFM net short position reports"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32",
    "pandas>=2.2",
    "apscheduler>=3.10",
    "pytz>=2024.1",
    "tenacity>=9.0",
    "streamlit>=1.35",
    "altair>=5.3",
    "python-dateutil>=2.9",
]

[project.scripts]
afm-scraper = "scraper.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scraper", "src/analytics"]

[tool.pytest.ini_options]
pythonpath = ["src"]
